--- a/resources/scripts/app.ts
+++ b/resources/scripts/app.ts
@@ -15,8 +15,3 @@ domReady(async () => {
     },
   });
 });
-
-/**
- * @see {@link https://webpack.org/api/hot-module-replacement/}
- */
-import.meta.webpackHot?.accept(console.error); // eslint-disable-line no-console