--- a/resources/scripts/app.ts
+++ b/resources/scripts/app.ts
@@ -1,17 +1,12 @@
-import domReady from '@roots/sage/client/dom-ready';
 import { createApp } from 'vue';
 import { NMenu } from 'naive-ui';
-/**
- * Application entrypointé
- */
-domReady(async () => {
-  const nav = createApp({
-    components: {
-      'n-menu': NMenu,
-    },
-    setup() {
-      console.log('teuds');
-      const menuItems = [];
-    },
-  });
+
+createApp({
+  components: {
+    'n-menu': NMenu,
+  },
+  setup() {
+    console.log('teuds');
+    const menuItems = [];
+  },
 });